+++ src/rj_strategy/include/rj_strategy/agent/position/defense.hpp
@@ -15,15 +15,8 @@
 #include <rj_geometry/point.hpp>
 #include <rj_msgs/action/robot_move.hpp>
 
-#include "planning/instant.hpp"
-#include "position.hpp"
-#include "rj_common/field_dimensions.hpp"
-#include "rj_common/time.hpp"
-#include "rj_constants/constants.hpp"
-#include "rj_geometry/geometry_conversions.hpp"
-#include "rj_geometry/point.hpp"
-#include "waller.hpp"
-
+#include "rj_strategy/agent/position.hpp"
+#include "rj_strategy/agent/position/waller.hpp"
 namespace strategy {
 
 /*
