+++ src/rj_strategy/include/rj_strategy/coordinator/marking_client.hpp
@@ -3,6 +3,7 @@
 #include <functional>
 
 #include <rclcpp/rclcpp.hpp>
+#include <spdlog/spdlog.h>
 
 #include <rj_msgs/msg/marking.hpp>
 #include <rj_msgs/srv/marking.hpp>
